Global update.
This project is the first one I created in Go to learn more about
the core concepts of the languages. It still alive out there but
fairly outdated, it is time to lift it up a little.
This PR doesn't include a full refact but drops useless dep and
version.